Tutorial 2: Updated Model
In order to continue to support the hands NOT jumping wildly at the `359->0` transition, I've made some updates to reduce the amount of processing that actually needs to be executed on each iteration.

The significant downside of this approach here is that it uses counters and not the _actual_ time on each iteration. This means that if the computer goes to sleep or processing slows, etc, the clock will no longer be accurate as the real time is never again checked after page load.